requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
protobuf==6.30.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
triton==2.3.0
typing_extensions==4.13.2
urllib3==2.4.0
datasets~=3.6.0
scikit-learn==1.6.1
evaluate==0.4.3
jiwer==3.1.0